Prémade users in db not readable through api
I finished writing the most important features of the api. Now the issue is that there is a very strange bug were the api can only read users that were created before the server was able to read the file and is not able to read any auther users data that has been added later on.

If we go through the register page to add a new user then it works and we can authenticate as that user. But it is impossible for the api to read the rest of the valid users. We can see them but their credentials are not valid or not considered authorized users.