could not fix issue
tried multiple debug methods and solution but nothing worked. The database cannot be read by the api if it has been modified directly in the json file. Decided to stop troubleshooting and just remove the data that was unusable. Still able to register new users and use those in the database